Refactor App component into separate file to prepare for Redux
